src/backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community  # 确保有这个，处理工具集
firecrawl-py
python-dotenv
supabase
pydantic
requests             # Serper 通常直接用 requests 调用 APItavily-python
exa-py
